Pulling ketoconnect's parse_title into separate function with its own tests
